% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-haplotypes.R
\name{cast.haplotypes}
\alias{cast.haplotypes}
\title{Cast the cleaned, phased, segmented data into a new shape, split haplotypes}
\usage{
cast.haplotypes(segs)
}
\arguments{
\item{segs}{matrix of segments with their class}
}
\value{
a matrix where the rows are cells and the columns are recurrent segments for each haplotype
}
\description{
Cast the cleaned, phased, segmented data into a new shape, split haplotypes
}
\author{
Alex van Vliet
}
