% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-majority.R
\name{cast.majority}
\alias{cast.majority}
\title{Cast the phased majority states into a new shape, split haplotypes}
\usage{
cast.majority(segs)
}
\arguments{
\item{segs}{matrix of sample + cell + chromosomes with their majority class}
}
\value{
a matrix where the rows are cells and the columns are haplotypes
}
\description{
Cast the phased majority states into a new shape, split haplotypes
}
\author{
Alex van Vliet
}
