% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincleaning.R
\name{clean.bins}
\alias{clean.bins}
\title{Bin cleaning: remove blacklisted bins and None bins}
\usage{
clean.bins(counts, blacklist)
}
\arguments{
\item{blacklist}{bin blacklist for centromeres and adjacent hard-to-sequence regions}

\item{count.data}{count matrix for the current sample}
}
\value{
the count matrix with centromeric bins and bins with no reads (None bins) removed
}
\description{
Bin cleaning: remove blacklisted bins and None bins
}
\author{
Alex van Vliet
}
