% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-outliers.R
\name{get.outliers}
\alias{get.outliers}
\title{Get all cells that do not conform to a translocation pattern to check for SVs/SCEs}
\usage{
get.outliers(phased, hapmatrix, translocations, x)
}
\arguments{
\item{phased}{the StrandphaseR file with all phased segments}

\item{hapmatrix}{the matrix that contains the split haplotypes for each chromosome/segment}

\item{translocations}{the file with all potential translocations}

\item{x}{the rownumber of the translocations file}
}
\description{
Get all cells that do not conform to a translocation pattern to check for SVs/SCEs
}
\author{
Alex van Vliet
}
