% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethaplopq.R
\name{gethaplopq}
\alias{gethaplopq}
\title{Split haplotypes per chromosome arm}
\usage{
gethaplopq(phased)
}
\arguments{
\item{phased}{MosaiCatcher phased segments file}
}
\value{
data.table with all haplotype/arm combinations
}
\description{
Split haplotypes per chromosome arm
}
\author{
Alex van Vliet
}
