% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasecounts.R
\name{phase.counts}
\alias{phase.counts}
\title{Phase MosaiCatcher count data using StrandPhaseR output}
\usage{
phase.counts(counts, phased.segments)
}
\arguments{
\item{counts}{clean count matrix}

\item{phased.segments}{StrandPhaseR output}
}
\value{
count matrix with WC -> CW as computed by StrandPhaseR
}
\description{
Phase MosaiCatcher count data using StrandPhaseR output
}
\author{
Alex van Vliet
}
