% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gethaplopq.R, R/pq-cast.R
\name{pq.cast}
\alias{pq.cast}
\alias{pq.cast}
\title{Split haplotypes per chromosome arm}
\usage{
pq.cast(phased)

pq.cast(phased)
}
\arguments{
\item{phased}{MosaiCatcher phased segments file}

\item{phased}{MosaiCatcher phased segments file}
}
\value{
data.table with all haplotype/arm combinations

data.table with all haplotype/arm combinations
}
\description{
Split haplotypes per chromosome arm

Split haplotypes per chromosome arm
}
\author{
Alex van Vliet

Alex van Vliet
}
