% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrentsegments.R
\name{rec.seg}
\alias{rec.seg}
\title{Get recurrent segments}
\usage{
rec.seg(counts.seg)
}
\arguments{
\item{counts.seg}{segmented count matrix (segfinder.R output)}
}
\value{
a list of recurrent segments that are not whole chromosomes and at least 1Mb long
}
\description{
Get recurrent segments
}
\author{
Alex van Vliet
}
