% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region-to-tr.R
\name{region.to.tr}
\alias{region.to.tr}
\title{From a list of regions get the corresponding strand state for that region in each cell}
\usage{
region.to.tr(regions, phased)
}
\arguments{
\item{regions}{all regions for the current sample}

\item{phased}{the phased MosaiCatcher data}
}
\value{
a matrix with as many columns as regions. Each row is a cell, each column is the strand state for each cell in that region.
}
\description{
From a list of regions get the corresponding strand state for that region in each cell
}
\author{
Alex van Vliet
}
