% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segfinder.R
\name{seg.finder}
\alias{seg.finder}
\title{Segment finder for TranslocatoR}
\usage{
seg.finder(m)
}
\arguments{
\item{m}{count matrix or segment matrix with sample, cell, chrom, start, end and class}
}
\value{
a list of breakpoints where strand state changes take place
}
\description{
Segment finder for TranslocatoR
}
\author{
Alex van Vliet
}
