% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentcheck.R
\name{segment.check}
\alias{segment.check}
\title{Edit Mosaicatcher segments if necessary}
\usage{
segment.check(mc.segments, rec.segments)
}
\arguments{
\item{mc.segments}{segments as determined by MosaiCatcher (with lowest sse)}

\item{rec.segments}{recurrent segemnts determined from count data}
}
\value{
mc.segments with new segments added if necessary
}
\description{
Edit Mosaicatcher segments if necessary
}
\author{
Alex van Vliet
}
