% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translocator.R
\name{translocatoR}
\alias{translocatoR}
\title{Find translocations}
\usage{
translocatoR(data.folder, output.folder, samples, options = "segments",
  binsize = 100000L, cutoff = 0.01, regions = NULL, trfile = NULL,
  blacklist = T)
}
\arguments{
\item{data.folder}{path to MosaiCatcher data folder}

\item{output.folder}{absolute path to output folder for TranslocatoR data}

\item{samples}{on which samples in the MosaiCatcher output folder would you like to run TranslocatoR}

\item{options}{can take values "segments", "pq", "majority". Defaults to "segments".}

\item{binsize}{which binsize to use, defaults to 100kb}

\item{cutoff}{cutoff for significant FDR-corrected p-values, defaults to 0.01}

\item{regions}{list of regions in the format "chr#:<start>-<end>" for potential translocations}

\item{trfile}{path to txt file with a suspected translocation. 
Can contain multiple translocations for one sample in separate columns, 
can also be a list of files for different samples. 
The first two columns must contain the cell IDs and the sample name.}

\item{blacklist}{whether to use the blacklist for centromeres and short arms for acrocentric chromosomes. defaults to True. use is strongly recommended.}
}
\value{
matrix
}
\description{
Find translocated chromosomal segments by correlating haplotype strand states in Mosaicatcher output. 
Fully vectorized approach heavily reliant on data.table.
}
\examples{
translocatoR(data.folder = "/g/korbel2/StrandSeq/20180727_MosiacatcherResults/all-2018-08-02", output.folder = "/g/korbel/vliet/translocations", samples = c("RPE-BM510", "C7-data"))
}
\author{
Alex van Vliet
}
